flatten namespace, windows module, deprecate old calls
