+++ tests/test_tool.py
@@ -6,7 +6,7 @@ except ImportError:
     plt = None
 
 import rasterio
-from rasterio.tool import show, show_hist, stats
+from rasterio.rio.insp import show, show_hist, stats
 
 def test_stats():
     with rasterio.drivers():
